refactor: cli packages structure
