+++ cmd/cli/kubectl-kyverno/commands/command.go
@@ -9,7 +9,7 @@ import (
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/oci"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/version"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/experimental"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/experimental"
 	"github.com/spf13/cobra"
 )
