+++ cmd/cli/kubectl-kyverno/commands/test/test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/go-git/go-billy/v5"
 	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	"github.com/kyverno/kyverno/api/kyverno/v1beta1"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/filter"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/common"
-	filterutils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/filter"
 	pathutils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/path"
 	sanitizederror "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/sanitizedError"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/store"
@@ -34,7 +34,7 @@ func applyPoliciesFromPath(
 	policyResourcePath string,
 	rc *resultCounts,
 	openApiManager openapi.Manager,
-	filter filterutils.Filter,
+	filter filter.Filter,
 	auditWarn bool,
 ) ([]api.TestResults, []engineapi.EngineResponse, error) {
 	engineResponses := make([]engineapi.EngineResponse, 0)
