+++ cmd/cli/kubectl-kyverno/output/color/color.go
