+++ cmd/cli/kubectl-kyverno/output/color/color_test.go
