+++ cmd/cli/kubectl-kyverno/test/filter/filter_test.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/api"
 )
 
 func Test_policy_Apply(t *testing.T) {
