+++ cmd/cli/kubectl-kyverno/test/load_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	policyreportv1alpha2 "github.com/kyverno/kyverno/api/policyreport/v1alpha2"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/api"
 )
 
 func TestTestCases_Errors(t *testing.T) {
